Set the BSS color on the 5GHz radio
BSS coloring seems to default to 128. Anything above 63 causes the
OpenWrt to select a color at random. There are reports of this not
working correctly[1] so set this manually to a sensible number.

At some point OpenWrt will have support for color collision detection
and random color selection, at which point this can go.

1. https://forum.openwrt.org/t/belkin-rt3200-linksys-e8450-wifi-ax-discussion/94302/1006?page=117