Use channel 132 on the 5GHz band and be explicit about htmodes
In the UK it is permitted to transmit at the higher power of 26dBm on
channels 100-144. This theoretically allows a greater range. These
channels are sometimes used by radar stations and require DFS to select
different channels if a radar signal is detected. However, that is not
relevant in my case and these are all safe to use.

A wifi scan shows that all these channels are free. I want to use a
80MHz wide channel on the 5GHz band and so using channel 132 allows the
use of 5660-5720MHz, all of which are permitted at the higher power of
26 dBm.

The 2.4GHz band is pretty congested here so keep auto channel selection
for that.

Set the htmode to HT20 on radio0 for a 20MHz wide N channel on the
2.4GHz band and HE80 on radio1 for an 80MHz wide AX channel on the 5GHz
band. These are the defaults but it's worth being explicit.